cleaning up finalizing code - termcolor!

+++ builder/build.py
@@ -10,7 +10,7 @@ import shutil
 import datetime
 import concurrent.futures
 import multiprocessing
-  
+from termcolor import colored
 
 class SubLogger:
   def __init__(self, logger, name):
@@ -25,6 +25,7 @@ class SubLogger:
     self.lines.append(f'[{self.name}] - INFO - {line}')
 
   def error(self, line):
+    print(colored(line, 'red'))
     self.lines.append(f'[{self.name}] - ERROR - {line}')
 
 class MyWeirdLogger:
@@ -51,23 +52,22 @@ sys.path.append('..')
 from shared import download_otp
 from shared.download_otp import CURRENT_OTP_JAR
 
-
-globalLogger = MyWeirdLogger('feed-import-%s.log' % datetime.datetime.now().strftime('%Y-%m-%d-%H-%M'))
+globalLogger = None
 
 def checkForTransitFeed():
   # check if we've checked out transitfeed
   if not os.path.isdir('transitfeed'):
-    print("Don't have transitfeed, checkout out from git")
-    subprocess.run(["git", "clone", "https://github.com/google/transitfeed"])  # doesn't capture output
+    print('Don't have transitfeed, checkout out from git')
+    subprocess.run(['git', 'clone', 'https://github.com/google/transitfeed'])  # doesn't capture output
   else:
-    print("great, you have transitfeed")
+    print('great, you have transitfeed')
 
 def checkForPython2():
   try:
-    subprocess.run("python2 --version", shell=True)
-    print("great, you have python2")
+    subprocess.run('python2 --version', shell=True)
+    print('great, you have python2')
   except:
-    print("transitfeed requires python2, stupid thing, which you don't have. gotta go get that. exiting")
+    print(colored('transitfeed requires python2, stupid thing, which you don't have. gotta go get that. exiting', 'red'))
     sys.exit(1)
 
 def checkDirectories():
@@ -136,6 +136,7 @@ def downloadAndCheckFeedVersion_TransitFeeds(currentVersion, logger):
       logger.info(f'downloaded fileDest')
     except:
       traceback.print_exc()
+      print(colored(f'could not download {url}', 'red'))
       return MISSING_VALUE
 
   # verify feed
@@ -166,33 +167,38 @@ def downloadAndCheckFeedVersion_TransitFeeds(currentVersion, logger):
   return (True, fileDest, id)
 
 def downloadAndCheckFeed_TransitFeeds_FeedId(feedId):
-  logger = globalLogger.getWriter('transitfeeds:%s' % feedId)
+  try:
+    logger = globalLogger.getWriter('transitfeeds:%s' % feedId)
 
-  feedResponse = transitfeeds.getFeedVersions(feedId)
-  
-  versions = feedResponse['results']['versions']
-  title = ''
-  
-  if len(versions) == 0:
-    logger.info(f'No feeds at all for {feedId}')
+    feedResponse = transitfeeds.getFeedVersions(feedId)
+    
+    versions = feedResponse['results']['versions']
+    title = ''
+    
+    if len(versions) == 0:
+      logger.error(f'No feeds at all for {feedId}')
+      return False
+    
+    for index, currentVersion in enumerate(versions):
+      if not title:
+        print('processing %s' % currentVersion['f']['t'])
+      title = currentVersion['f']['t']
+      (status, filepath, id) = downloadAndCheckFeedVersion_TransitFeeds(currentVersion, logger)
+      if status:
+        destFilename = id.replace('/', '-') + '-gtfs.zip'
+        destPath = os.path.join(OTP_INPUT_DIR, destFilename)
+        logger.info(f'got good version {id}, copying {filepath} to {destPath}')
+        shutil.copyfile(filepath, destPath)
+        return True
+      if index > 5:
+        logger.error(f'Not looking back more than five versions for {feedId}')
+
+    print('feedid3')
+    logger.error(f'No good versions for {feedId} - {title}')
     return False
-  
-  for index, currentVersion in enumerate(versions):
-    if not title:
-      print('processing %s' % currentVersion['f']['t'])
-    title = currentVersion['f']['t']
-    (status, filepath, id) = downloadAndCheckFeedVersion_TransitFeeds(currentVersion, logger)
-    if status:
-      destFilename = id.replace('/', '-') + '-gtfs.zip'
-      destPath = os.path.join(OTP_INPUT_DIR, destFilename)
-      logger.info(f'got good version {id}, copying {filepath} to {destPath}')
-      shutil.copyfile(filepath, destPath)
-      return True
-    if index > 5:
-      logger.error(f'Not looking back more than five versions for {feedId}')
-
-  logger.error(f'No good versions for {feedId} - {title}')
-  return False
+  except:
+    traceback.print_exc()
+    
 
 def downloadAndCheckFeed_TransitFeeds_Location(location):
   feeds = transitfeeds.getLocation(location)
@@ -205,7 +211,7 @@ def downloadAndCheckFeed_TransitFeeds(feed):
   elif 'id' in feed:
     downloadAndCheckFeed_TransitFeeds_FeedId(feed['id'])
   else:
-    print(f"don't know how to parse transitfeeds feed {feed}")
+    print(colored(f'don't know how to parse transitfeeds feed {feed}', 'red'))
     sys.exit(1)
 
 def checkOTP():
@@ -220,38 +226,56 @@ def downloadFeeds():
   with concurrent.futures.ThreadPoolExecutor(max_workers=5) as executor:
     executor.map(downloadFeed, get_feeds()['feeds'])
 
-def runOTP():
+def runOTP(otpInputDir, otpOutputDir):
   otpJarName = CURRENT_OTP_JAR.rsplit('/', 1)[-1]
-  subprocess.run(["java", "-Xmx10G", "-jar", otpJarName, '--build', OTP_INPUT_DIR])
-  GRAPH_FILE = 'Graph.obj'
-  graphFilePath = os.path.join(OTP_INPUT_DIR, GRAPH_FILE)
+  completedProcess = subprocess.run(['java', '-Xmx10G', '-jar', otpJarName, '--build', otpInputDir])
+  if completedProcess.returncode != 0:
+    print(colored('****************', 'red'))
+    print(colored('OTP BUILD FAILED', 'red'))
+    print(colored('****************', 'red'))
+    sys.exit(1)
+
+  GraphFileName = 'Graph.obj'
+  graphFilePath = os.path.join(otpInputDir, GraphFileName)
   if os.path.exists(graphFilePath):
     print(f'built index at {graphFilePath}')
 
-    # copy graph from OTP_INPUT_DIR/Graph.obj to OTP_OUTPUT_DIR/date_str/Graph.obj
+    # copy graph from otpInputDir/Graph.obj to OTP_OUTPUT_DIR/date_str/Graph.obj
     dateStr = datetime.datetime.now().strftime('%Y-%m-%d-%H-%M')
-    outputDir = os.path.join(OTP_OUTPUT_DIR, dateStr)
-    outputPath = os.path.join(outputDir, GRAPH_FILE)
+    outputDir = os.path.join(otpOutputDir, dateStr)
+    outputPath = os.path.join(outputDir, GraphFileName)
     os.mkdir(outputDir)
     shutil.copy(graphFilePath, outputPath)
 
+    print(colored(f'All done at {outputPath}', 'green'))
     # symlink OTP_OUTPUT_DIR/datestr/ to OTP_OUTPUT_DIR/default/
-    currentLinkPath = os.path.join(OTP_OUTPUT_DIR, DEFAULT_GRAPH_NAME)
-    if os.path.islink(currentLinkPath):
-      os.unlink(currentLinkPath)
-    os.symlink(dateStr, currentLinkPath)
-
-    print(f'All done at {outputPath}')
-    print(f'linking to {currentLinkPath}')
-    print(f'To check by hand run')
-    print(f'java -Xmx2G -jar {CURRENT_OTP_JAR} --router default --graphs graphs --server')
+    currentLinkPath = os.path.join(otpOutputDir, DEFAULT_GRAPH_NAME)
+    if os.path.exists(currentLinkPath):
+      if os.path.islink(currentLinkPath):
+        os.unlink(currentLinkPath)
+        os.symlink(dateStr, currentLinkPath)
+        print(f'linking to {currentLinkPath}')
+        print(f'To check by hand run')
+        print(f'  java -Xmx2G -jar {CURRENT_OTP_JAR} --router default --graphs graphs --server')
+      else:
+        print(colored(f'{currentLinkPath} is not a symlink, not removing and not updating', 'yellow'))
+        print(f'To check by hand run')
+        print(f'  java -Xmx2G -jar {CURRENT_OTP_JAR} --router {dateStr} --graphs graphs --server')
+    
     print('check that it is working in nyc with')
     print('http://localhost:8080/otp/routers/default/plan?fromPlace=40.700853,-73.947738&toPlace=40.741524,%20-73.989330')
 
+def main():
+  # if we do this in the global scope there's a weird multiprocessing error, sorry, gross
+  global globalLogger
+  globalLogger = MyWeirdLogger('feed-import-%s.log' % datetime.datetime.now().strftime('%Y-%m-%d-%H-%M'))
 
-checkForTransitFeed()
-checkForPython2()
-checkDirectories()
-checkOTP()
-downloadFeeds()
-runOTP()
+  checkForTransitFeed()
+  checkForPython2()
+  checkDirectories()
+  checkOTP()
+  downloadFeeds()
+  # runOTP(OTP_INPUT_DIR, OTP_OUTPUT_DIR)
+
+if __name__ == '__main__':
+    main()
